+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Taproot and Schnorr activated
 - Encrypted P2P transport (BIP324)
 - BLS aggregate staking
+- Dandelion++ transaction relay
 
 ## v2.0.2 (2024-11-12)
 - Working Transaction Hotfix: Displays correct Timestamp in Explorer